Delete on confirm small style changes

+++ src/modules/IdeaList/components/Idea/index.js
@@ -1,54 +1,59 @@
+import * as firebase from 'firebase';
 import { connect } from 'react-redux';
-import { getIdeaList, remove, update } from '../../actions';
+import { getIdeaList, remove, update, likeIdea } from '../../actions';
 import React, { Component, PropTypes } from 'react';
 
 class Idea extends Component {
 
   onClickRemove = () => {
+    if (!confirm("Are you sure you want to delete " + this.props.text + " ?")) {
+      return;
+    }
     const { onRemove } = this.props;
     const key = this.props.id;
     onRemove(key);
-    // this.props.onGetIdeaList();
-  }
-
-  // onClickLike = () => {
-  //   const { onLike } = this.props;
-  //   const key = this.props.id;
-  //   onLike(key, this.props.liked);
-  //   // this.props.onGetIdeaList();
-  // }
+  };
 
   onClickUpdateItem = () => {
     const key = this.props.id;
     const { onUpdateItem } = this.props;
     const text = prompt("Edit task", this.props.text);
-    this.props.onUpdateItem(key, text);
-    // this.props.onGetIdeaList();
-  }
+    onUpdateItem(key, text);
+  };
+
+  onClickLikeIdea = () => {
+    const key = this.props.id;
+    const { onLikeIdea } = this.props;
+    onLikeIdea(key);
+
+  };
 
 
   render() {
-    const{
+    const {
       onClickRemove,
-      // onClickLike,
       onClickUpdateItem,
-      props : {
-        likes,
-        text,
-        id,
-        editable,
-    }} = this;
+      onClickLikeIdea,
+      props: {
+
+      }
+    } = this;
+
+    const userIsOwner = this.props.user === firebase.auth().currentUser.uid;
 
     return (
       <div>
         <div className="idea-item">
-          <span className="like-mark">&#9899; {this.props.likes}</span>
+          <span className="heart-mark /*liked or unliked*/">{this.props.likes} &#10084;</span>
           <span>{this.props.text}</span>
-          { editable &&
-            <button className="x-mark" onClick={onClickRemove}>REMOVE</button>
+          {userIsOwner &&
+          <button className="x-mark" onClick={onClickRemove}>REMOVE</button>
           }
-          { editable &&
-            <button onClick={onClickUpdateItem} className="edit-mark">EDIT</button>
+          {userIsOwner &&
+          <button onClick={onClickUpdateItem} className="edit-mark">EDIT</button>
+          }
+          {!userIsOwner &&
+          <button onClick={onClickLikeIdea} className="like-mark">LIKE</button>
           }
         </div>
       </div>
@@ -58,22 +63,23 @@ class Idea extends Component {
 
 const mapStateToProps = state => ({
   ideaList: state.idea.ideaList,
-})
+});
 
 const mapDispatchToProps = dispatch => ({
-  onUpdateItem(key, newText){
-    dispatch(update(key, {text: newText}));
+  onUpdateItem(key, newText) {
+    dispatch(update(key, { text: newText }));
   },
 
-  onRemove(key){
+  onRemove(key) {
     dispatch(remove(key));
   },
-  // onLike(key, isLiked){
-  //   dispatch(update(key, {liked: !isLiked}));
-  // },
-  onGetIdeaList(){
+  onGetIdeaList() {
     dispatch(getIdeaList());
   },
+  onLikeIdea(key) {
+    dispatch(likeIdea(key));
+  },
+
   dispatch,
 });
 
@@ -81,9 +87,9 @@ Idea.propTypes = {
   likes: PropTypes.number.isRequired,
   text: PropTypes.string.isRequired,
   onRemove: PropTypes.func.isRequired,
-  editable: PropTypes.bool.isRequired,
+  user: PropTypes.string.isRequired,
   onUpdateItem: PropTypes.func.isRequired,
-  // onLike: PropTypes.func.isRequired,
+  onLikeIdea: PropTypes.func.isRequired,
 };
 
 export default connect(mapStateToProps, mapDispatchToProps)(Idea);
